chore: writing FAT12 entries no longer reads sectors ahead-of-time
When writing a FATEntry on a FAT12 filesystem, the bytes the entry spans would be read ahead-of-time & OR-ed to the new entry's bytes. This is no longer the case. FAT12 still reads those bytes since FAT12 entries are 1.5 bytes long, but now they are read individually just before the new byte is written.