chore: finalize DirEntry revamp
From my testing, the user should now be able to have 2 or more directories or files open in a RO manner simultaneously without any issues. [#14](https://github.com/Oakchris1955/simple-fatfs/issues/14) mentions that there could be data corruption if the same happens with a RW file / an RO file is open and a RW operation deletes the file / its parent directory, but that is out of the scope of this commit.

Updated testing accordingly

Minor changes to ReadDir and ReadDirInt.